backend/requirements.txt
# backend/requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic-settings